Interpretador com depuração
