+++ fontes/interpretador/interpretador-com-depuracao.ts
@@ -159,7 +159,7 @@ export class InterpretadorComDepuracao
         }
     }
 
-    protected async avaliarArgumentosEscreva(argumentos: Construto[]): Promise<string> {
+    async avaliarArgumentosEscreva(argumentos: Construto[]): Promise<string> {
         let formatoTexto: string = '';
 
         for (const argumento of argumentos) {
